+++ packages/select-with-tags/src/components/tag/component.tsx
@@ -1,12 +1,16 @@
 import React, { useCallback } from 'react';
+import cn from 'classnames';
 import { Tag as CoreTag } from '@alfalab/core-components-tag';
 import { CrossCompactMIcon } from '@alfalab/icons-glyph/CrossCompactMIcon';
-
 import { TagComponent } from '../../types';
-
 import styles from './index.module.css';
 
-export const Tag: TagComponent = ({ option: { content, key }, handleDeleteTag }) => {
+export const Tag: TagComponent = ({
+    option: { content, key },
+    onClick,
+    handleDeleteTag,
+    ...props
+}) => {
     const handleClick = useCallback(() => {
         if (handleDeleteTag) {
             handleDeleteTag(key);
@@ -14,11 +18,19 @@ export const Tag: TagComponent = ({ option: { content, key }, handleDeleteTag })
     }, [handleDeleteTag, key]);
 
     return (
-        <CoreTag key={key} size='xs' checked={true} className={styles.tag}>
+        <CoreTag
+            key={key}
+            size='xs'
+            onClick={onClick}
+            checked={!!handleDeleteTag}
+            className={cn(styles.tag, { [styles.tagNoClose]: !handleDeleteTag })}
+            {...props}
+        >
             <span className={styles.tagContentWrap}>
                 {content}
-
-                <CrossCompactMIcon onClick={handleClick} className={styles.tagCross} />
+                {handleDeleteTag && (
+                    <CrossCompactMIcon onClick={handleClick} className={styles.tagCross} />
+                )}
             </span>
         </CoreTag>
     );