+++ packages/dark-theme-styles-injector/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-dark-theme-styles-injector@1.4.4...@alfalab/core-components-dark-theme-styles-injector@1.4.5) (2021-02-19)
+
+**Note:** Version bump only for package @alfalab/core-components-dark-theme-styles-injector
+
+
+
+
+
+## [1.4.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-dark-theme-styles-injector@1.4.3...@alfalab/core-components-dark-theme-styles-injector@1.4.4) (2021-02-19)
+
+**Note:** Version bump only for package @alfalab/core-components-dark-theme-styles-injector
+
+
+
+
+
 ## [1.4.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-dark-theme-styles-injector@1.4.2...@alfalab/core-components-dark-theme-styles-injector@1.4.3) (2021-02-19)
 
 **Note:** Version bump only for package @alfalab/core-components-dark-theme-styles-injector