+++ packages/select/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.3.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-select@6.3.1...@alfalab/core-components-select@6.3.2) (2021-02-19)
+
+**Note:** Version bump only for package @alfalab/core-components-select
+
+
+
+
+
+## [6.3.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-select@6.3.0...@alfalab/core-components-select@6.3.1) (2021-02-19)
+
+
+### Bug Fixes
+
+* **select:** min width bug ([#524](https://github.com/alfa-laboratory/core-components/issues/524)) ([cdd686b](https://github.com/alfa-laboratory/core-components/commit/cdd686ba0780fdde982fff6629afd2fa798f9479))
+
+
+
+
+
 # [6.3.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-select@6.2.1...@alfalab/core-components-select@6.3.0) (2021-02-18)